Fixed type assignment error.
